# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  online: 'online',
  away: 'away',
  offline: 'offline',
} as const;

export interface User {
  id: string;
  name: string;
  initials: string;
  avatarColor?: string;
  role?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  needsOnboarding: boolean;
  status: UserStatus;
  /** @nullable */
  lastSeen?: string | null;
}

export interface ProfileInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  /**
     * @minLength 10
     * @maxLength 10
     */
  dateOfBirth: string;
}

export type PresenceStatus = typeof PresenceStatus[keyof typeof PresenceStatus];


export const PresenceStatus = {
  online: 'online',
  away: 'away',
  offline: 'offline',
} as const;

export interface Presence {
  userId: string;
  status: PresenceStatus;
  /** @nullable */
  lastSeen?: string | null;
}

export type ConversationKind = typeof ConversationKind[keyof typeof ConversationKind];


export const ConversationKind = {
  direct: 'direct',
  group: 'group',
} as const;

export type MessageStatus = typeof MessageStatus[keyof typeof MessageStatus];


export const MessageStatus = {
  sent: 'sent',
  delivered: 'delivered',
  read: 'read',
} as const;

export interface Message {
  id: string;
  conversationId: string;
  senderId: string;
  senderName: string;
  senderInitials?: string;
  body: string;
  sentAt: string;
  status: MessageStatus;
  replyTo?: Message | null;
}

export interface Conversation {
  id: string;
  name: string;
  kind: ConversationKind;
  avatarColor?: string;
  participants: User[];
  unreadCount: number;
  pinned?: boolean;
  muted?: boolean;
  lastMessage: Message | null;
}

export type ConversationInputKind = typeof ConversationInputKind[keyof typeof ConversationInputKind];


export const ConversationInputKind = {
  direct: 'direct',
  group: 'group',
} as const;

export interface ConversationInput {
  participantIds: string[];
  name?: string;
  kind?: ConversationInputKind;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  body: string;
}

export type ListUsersParams = {
search?: string;
};