# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Chat and messaging operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /profile:
    get:
      operationId: getProfile
      tags: [chat]
      summary: Get the current demo profile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateProfile
      tags: [chat]
      summary: Complete or update the current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [chat]
      summary: Search users
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /presence:
    get:
      operationId: listPresence
      tags: [chat]
      summary: Get current online presence
      responses:
        "200":
          description: Presence entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Presence"
  /conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List conversations for the current user
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create a direct or group conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    User:
      type: object
      required: [id, name, initials, status, needsOnboarding]
      properties:
        id: { type: string }
        name: { type: string }
        initials: { type: string }
        avatarColor: { type: string }
        role: { type: string }
        email: { type: ["string", "null"] }
        dateOfBirth: { type: ["string", "null"] }
        needsOnboarding: { type: boolean }
        status: { type: string, enum: [online, away, offline] }
        lastSeen: { type: ["string", "null"] }
    ProfileInput:
      type: object
      required: [name, dateOfBirth]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        dateOfBirth: { type: string, minLength: 10, maxLength: 10 }
    Presence:
      type: object
      required: [userId, status]
      properties:
        userId: { type: string }
        status: { type: string, enum: [online, away, offline] }
        lastSeen: { type: ["string", "null"] }
    Conversation:
      type: object
      required: [id, name, kind, participants, unreadCount, lastMessage]
      properties:
        id: { type: string }
        name: { type: string }
        kind: { type: string, enum: [direct, group] }
        avatarColor: { type: string }
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        unreadCount: { type: integer }
        pinned: { type: boolean }
        muted: { type: boolean }
        lastMessage:
          anyOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
    Message:
      type: object
      required: [id, conversationId, senderId, senderName, body, sentAt, status]
      properties:
        id: { type: string }
        conversationId: { type: string }
        senderId: { type: string }
        senderName: { type: string }
        senderInitials: { type: string }
        body: { type: string }
        sentAt: { type: string }
        status: { type: string, enum: [sent, delivered, read] }
        replyTo:
          anyOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
    ConversationInput:
      type: object
      required: [participantIds]
      properties:
        participantIds:
          type: array
          items: { type: string }
        name: { type: string }
        kind: { type: string, enum: [direct, group] }
    MessageInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 4000 }