# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  /**
     * @minLength 10
     * @maxLength 10
     */
  dateOfBirth: string;
}