# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from './userStatus';

export interface User {
  id: string;
  name: string;
  initials: string;
  avatarColor?: string;
  role?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  needsOnboarding: boolean;
  status: UserStatus;
  /** @nullable */
  lastSeen?: string | null;
}